chore(zui): move transform utils
