+++ zui/src/transforms/zui-to-typescript-type/index.ts
@@ -5,7 +5,7 @@ import {
   toPropertyKey,
   toTypeArgumentName,
   primitiveToTypscriptLiteralType,
-} from './utils'
+} from '../common/utils'
 import * as errors from '../common/errors'
 
 const Primitives = [
